Rename UML.modelfatory to UML.recipes
This name covers the contents of the module better. Also it's less confusing,
since the terms model and factory are used with different meanings.
