+++ gaphor/RAAML/fta/dormantevent.py
@@ -14,7 +14,7 @@ from gaphor.diagram.support import represents
 from gaphor.diagram.text import FontStyle, FontWeight
 from gaphor.RAAML import raaml
 from gaphor.RAAML.fta.undevelopedevent import draw_undeveloped_event
-from gaphor.UML.modelfactory import stereotypes_str
+from gaphor.UML.recipes import stereotypes_str
 
 
 @represents(raaml.DormantEvent)
