+++ gaphor/RAAML/fta/transferout.py
@@ -15,7 +15,7 @@ from gaphor.diagram.text import FontStyle, FontWeight
 from gaphor.RAAML import raaml
 from gaphor.RAAML.fta.constants import DEFAULT_FTA_MAJOR
 from gaphor.RAAML.fta.transferin import draw_transfer_in
-from gaphor.UML.modelfactory import stereotypes_str
+from gaphor.UML.recipes import stereotypes_str
 
 
 @represents(raaml.TransferOut)
