+++ gaphor/SysML/blocks/proxyport.py
@@ -16,7 +16,7 @@ from gaphor.diagram.shapes import (
 )
 from gaphor.diagram.support import represents
 from gaphor.SysML import sysml
-from gaphor.UML.modelfactory import stereotypes_str
+from gaphor.UML.recipes import stereotypes_str
 
 
 def text_position(position):
