+++ gaphor/SysML/requirements/relationships.py
@@ -3,7 +3,7 @@ from gaphor.diagram.presentation import LinePresentation, Named
 from gaphor.diagram.shapes import Box, Text, draw_arrow_head
 from gaphor.diagram.support import represents
 from gaphor.SysML import sysml
-from gaphor.UML.modelfactory import stereotypes_str
+from gaphor.UML.recipes import stereotypes_str
 
 
 class DirectedRelationshipPropertyPathItem(LinePresentation, Named):
