+++ gaphor/UML/__init__.py
@@ -1,4 +1,4 @@
 # Here, order matters
 import gaphor.UML.umloverrides
-from gaphor.UML import modelfactory as model
+from gaphor.UML import recipes as model
 from gaphor.UML.uml import *
