+++ gaphor/UML/actions/activitynodes.py
@@ -15,7 +15,7 @@ from gaphor.diagram.presentation import ElementPresentation, HandlePositionUpdat
 from gaphor.diagram.shapes import Box, IconBox, Text, stroke
 from gaphor.diagram.support import represents
 from gaphor.i18n import gettext
-from gaphor.UML.modelfactory import stereotypes_str
+from gaphor.UML.recipes import stereotypes_str
 
 DEFAULT_JOIN_SPEC = "and"
