+++ gaphor/UML/interactions/interactionstoolbox.py
@@ -8,7 +8,7 @@ from gaphor.core import gettext
 from gaphor.diagram.diagramtoolbox import ToolDef, ToolSection, namespace_config
 from gaphor.diagram.tools import new_item_factory
 from gaphor.UML import diagramitems
-from gaphor.UML.modelfactory import owner_package
+from gaphor.UML.recipes import owner_package
 
 
 def interaction_config(new_item):
