+++ gaphor/UML/profiles/extensionconnect.py
@@ -1,8 +1,8 @@
 from gaphor import UML
 from gaphor.diagram.connectors import Connector, RelationshipConnect
 from gaphor.diagram.presentation import Classified
-from gaphor.UML.modelfactory import owner_package
 from gaphor.UML.profiles.extension import ExtensionItem
+from gaphor.UML.recipes import owner_package
 
 
 @Connector.register(Classified, ExtensionItem)
