+++ gaphor/UML/states/transition.py
@@ -4,7 +4,7 @@ from gaphor import UML
 from gaphor.diagram.presentation import LinePresentation, Named
 from gaphor.diagram.shapes import Box, Text, draw_arrow_tail
 from gaphor.diagram.support import represents
-from gaphor.UML.modelfactory import stereotypes_str
+from gaphor.UML.recipes import stereotypes_str
 
 
 @represents(UML.Transition)
