+++ gaphor/UML/tests/test_recipes.py
